Clearing first sprint

+++ сlearing_service/app/entities/TransactionDTO.scala
@@ -1,3 +1,3 @@
 package entities
 
-case class TransactionDTO(id: String, bankReceiver: String, bankSender: String, userReceiver: String, userSender: String, amount: Double)
+case class TransactionDTO(id: String, bankReceiver: String, bankSender: String, userReceiver: String, userSender: String, amount: Int)
